+++ Landscape/game_mod/adv_map.hpp
@@ -2,3 +2,5 @@
 
 
 void AdvMapPatch(class PatcherInstance& p);
+
+void AdvMapTownPatch(PatcherInstance& p);