+++ Landscape/game_mod/adv_mt.hpp
@@ -8,4 +8,6 @@ bool __fastcall IsTerrainType(const NewfullMap& map, const CObject& obj, int wid
 
 TerrainType8 __fastcall GetTerrainType(const NewfullMap& map, const CObject& obj, int width);
 
+TerrainType8 __fastcall GetTerrainType(const NewfullMap& map, const class town& tw);
+
 void __fastcall MarkMagicTerrain(NewfullMap& map);